refactor: tribute feature encapsulation
