+++ Source/Endpoints/Iwentys.Endpoint.Sdk/ControllerClients/Guilds/GuildTributeControllerClient.cs
@@ -2,7 +2,7 @@
 using System.Net.Http;
 using System.Net.Http.Json;
 using System.Threading.Tasks;
-using Iwentys.Features.Guilds.Models.GuildTribute;
+using Iwentys.Features.Tributes.Models;
 
 namespace Iwentys.Endpoint.Sdk.ControllerClients.Guilds
 {
