+++ Source/Endpoints/Iwentys.Endpoint.Sdk/Iwentys.Endpoint.Sdk.csproj
@@ -16,6 +16,7 @@
     <ProjectReference Include="..\..\Features\Iwentys.Features.Gamification\Iwentys.Features.Gamification.csproj" />
     <ProjectReference Include="..\..\Features\Iwentys.Features.Guilds\Iwentys.Features.Guilds.csproj" />
     <ProjectReference Include="..\..\Features\Iwentys.Features.Newsfeeds\Iwentys.Features.Newsfeeds.csproj" />
+    <ProjectReference Include="..\..\Features\Iwentys.Features.PeerReview\Iwentys.Features.Tributes.csproj" />
     <ProjectReference Include="..\..\Features\Iwentys.Features.Quests\Iwentys.Features.Quests.csproj" />
   </ItemGroup>
