+++ Source/Features/Iwentys.Features.PeerReview/Iwentys.Features.Tributes.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Iwentys.Features.Guilds\Iwentys.Features.Guilds.csproj" />
+  </ItemGroup>
+
+</Project>
