+++ Source/Features/Iwentys.Features.PeerReview/Models/ActiveTributeResponseDto.cs
@@ -1,8 +1,8 @@
 ﻿using System;
-using Iwentys.Features.Guilds.Entities;
-using Iwentys.Features.Guilds.Enums;
+using Iwentys.Features.Tributes.Entities;
+using Iwentys.Features.Tributes.Enums;
 
-namespace Iwentys.Features.Guilds.Models.Guilds
+namespace Iwentys.Features.Tributes.Models
 {
     public record ActiveTributeResponseDto
     {
