+++ Source/Features/Iwentys.Features.PeerReview/Models/TributeCompleteRequest.cs
@@ -1,4 +1,4 @@
-﻿namespace Iwentys.Features.Guilds.Models.GuildTribute
+﻿namespace Iwentys.Features.Tributes.Models
 {
     public class TributeCompleteRequest
     {
