+++ Source/Features/Iwentys.Features.PeerReview/Models/TributeInfoResponse.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Linq.Expressions;
-using Iwentys.Features.Guilds.Entities;
-using Iwentys.Features.Guilds.Enums;
 using Iwentys.Features.Students.Models;
+using Iwentys.Features.Tributes.Entities;
+using Iwentys.Features.Tributes.Enums;
 
-namespace Iwentys.Features.Guilds.Models.GuildTribute
+namespace Iwentys.Features.Tributes.Models
 {
     public class TributeInfoResponse
     {
