+++ Source/Iwentys.Database/Context/IwentysDbContext.cs
@@ -13,6 +13,7 @@ using Iwentys.Features.Newsfeeds.Entities;
 using Iwentys.Features.Quests.Entities;
 using Iwentys.Features.Students.Entities;
 using Iwentys.Features.Study.Entities;
+using Iwentys.Features.Tributes.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
