+++ Source/Iwentys.sln
@@ -66,6 +66,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Iwentys.Features.Tributes", "Features\Iwentys.Features.PeerReview\Iwentys.Features.Tributes.csproj", "{F5347FB0-396C-4F17-820F-8735E6F4E955}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -172,6 +174,10 @@ Global
 		{626DAF59-83C6-4939-87F5-0EEE60382476}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{626DAF59-83C6-4939-87F5-0EEE60382476}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{626DAF59-83C6-4939-87F5-0EEE60382476}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F5347FB0-396C-4F17-820F-8735E6F4E955}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F5347FB0-396C-4F17-820F-8735E6F4E955}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F5347FB0-396C-4F17-820F-8735E6F4E955}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F5347FB0-396C-4F17-820F-8735E6F4E955}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -198,6 +204,7 @@ Global
 		{607370AD-352D-4576-9757-BC27A3075DF2} = {70E1C0FF-AA6C-4ED5-AE29-587699CF170C}
 		{1194505C-80C5-4020-9260-3E1B766326CC} = {95512D3C-159E-4224-B79E-E445AE17B0E7}
 		{626DAF59-83C6-4939-87F5-0EEE60382476} = {95512D3C-159E-4224-B79E-E445AE17B0E7}
+		{F5347FB0-396C-4F17-820F-8735E6F4E955} = {95512D3C-159E-4224-B79E-E445AE17B0E7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C699E386-11E7-4ABB-9AD6-C676F3CA2061}
