+++ Source/Tests/Iwentys.Tests/Features/Guilds/GuildTributeServiceTest.cs
@@ -3,9 +3,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Iwentys.Features.GithubIntegration.Entities;
 using Iwentys.Features.Guilds.Models.Guilds;
-using Iwentys.Features.Guilds.Models.GuildTribute;
 using Iwentys.Features.Students.Domain;
 using Iwentys.Features.Students.Enums;
+using Iwentys.Features.Tributes.Models;
 using Iwentys.Tests.TestCaseContexts;
 using NUnit.Framework;
